requirements.txt
amqp==1.4.9
anyjson==0.3.3
Babel==2.5.3
billiard==3.3.0.23
celery==3.1.25
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
diff-match-patch==20121119

django-auth==0.1.9


django-celery==3.2.2
django-cors-headers==2.1.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9

django-redis==4.8.0
django-reversion==2.0.12

djangorestframework-jwt==1.11.0
docopt==0.6.2
drf-extensions==0.3.1
elasticsearch==6.3.1
et-xmlfile==1.0.1
flower==0.9.2
future==0.16.0
gunicorn==19.7.1
httplib2==0.10.3
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
kombu==3.0.37
Markdown==2.6.11
MarkupSafe==1.0
mysql-connector-python==8.0.6
mysqlclient==1.3.12
Naked==0.1.31
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.4.9
Pillow==5.0.0
pipreqs==0.4.9
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycrypto==2.6.1
PyJWT==1.5.3
PyMySQL==0.9.3
python-crontab==2.3.6
python-dateutil==2.7.3

python-utils==2.3.0
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
shellescape==3.4.1
six==1.11.0
tablib==0.12.1
tornado==4.5.3
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
utils==0.9.0
uuid==1.30
uWSGI==2.0.17.1
vine==1.1.4
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
yarg==0.1.9